Partially done with lasso/ridge assignment
getting some inconsistent behavior in the NewProstate.R file. sometimes
the looping works fine, other times it fails. No pattern thus far.
Thinking that it might have something to do with the cross-validation
sampling causing some kind of size error in the data. Going to continue
working, but hoping to resolve the issue in the future.